+++ lib/rumbleship-beeline.d.ts
@@ -11,10 +11,20 @@ export declare class RumbleshipBeeline {
     static initialize(configureBeeline: (config: HoneycombConfiguration) => any, config: HoneycombConfiguration): void;
     /**
      *
+     * @param server_like A Hapi.Server that has been instrumented.
+     *
+     * Replaces initial `RumbleshipBeeline` with  the beeline that has been
+     * reconfigured and extended by the `@hapi/hapi` instrumentation so `bindFunctionToTrace()`
+     * picks up the Hapi request context
+     */
+    static shimFromInstrumentation<T>(server_like: T): T;
+    /**
+     *
      * @param context_id The unique id for the context this beeline is operating in.
      * Likely `service_context_id` or `request_id`
      */
     static make(context_id: string): RumbleshipBeeline;
+    static flush(): any;
     constructor(context_id: string);
     /**
      *