+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+  * Expose `flush()` on our static RumbleshipBeeline
+  * RumbleshipBeeline.shimFromInstrumentation() to pull fully Hapi-instrumented beeline back into context factory
+### Removed
+### Changed
+### Fixed
+  * HoneycombMiddleware uses the fully typed RumbleshipBeeline
+### Deprecated
+### Security
+
+
 ## [1.1.0] -- 2020-04-15
 
 ### Changed